requirements.txt
amqp==5.3.1
appdirs==1.4.4
asgiref==3.8.1
billiard==4.2.1
black==21.7b0
cachetools==5.5.2
celery==5.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
dj-database-url==2.3.0
Django==5.1.7
django-extensions==3.2.3
django-simple-history==3.8.0
django-widget-tweaks==1.4.8
djangorestframework==3.16.0
Faker==36.1.1
flake8==3.9.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
isort==5.9.2
kombu==5.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==24.2
pathspec==0.8.1
pillow==10.3.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2021.1
redis==5.2.1
regex==2021.7.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sqlparse==0.4.4
tomli==1.0.4
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
